+++ hydra-tx/src/Hydra/Chain/ChainState.hs
@@ -18,6 +18,11 @@ newtype ChainSlot = ChainSlot Natural
 -- XXX: Find a better name for this. Maybe HasL1State, IsChainTx or IsL1Tx?
 class
   ( IsTx tx
+  , Ord (ChainPointType tx)
+  , Eq (ChainPointType tx)
+  , Show (ChainPointType tx)
+  , FromJSON (ChainPointType tx)
+  , ToJSON (ChainPointType tx)
   , Eq (ChainStateType tx)
   , Show (ChainStateType tx)
   , FromJSON (ChainStateType tx)
@@ -32,6 +37,13 @@ class
   -- XXX: Why is this not always UTxOType?
   type ChainStateType tx = c | c -> tx
 
-  -- | Get the chain slot for a chain state. NOTE: For any sequence of 'a'
-  -- encountered, we assume monotonically increasing slots.
-  chainStateSlot :: ChainStateType tx -> ChainSlot
+  -- | Get the chain point for a chain state.
+  chainStatePoint :: ChainStateType tx -> ChainPointType tx
+
+  -- | Get the chain slot from a chain point.
+  chainPointSlot :: ChainPointType tx -> ChainSlot
+
+-- | Get the chain slot for a chain state. NOTE: For any sequence of 'a'
+-- encountered, we assume monotonically increasing slots.
+chainStateSlot :: IsChainState tx => ChainStateType tx -> ChainSlot
+chainStateSlot = chainPointSlot . chainStatePoint