+++ CHANGELOG.md
@@ -24,7 +24,7 @@ changes.
   out of sync with the chain. Defaults to half the contestation period.
 
 - **BREAKING** Resume node from the latest observed tick.
-  - `TickObserved` event schema has changed: the `chainSlot` field has been replaced with `chainState`
+  - `TickObserved` event schema has changed: the `chainSlot` field has been replaced with `chainPoint`
   - `Greetings` message now also contains a new field `currentSlot` that indicates the last known slot while catching up.
   - See [Issue #2206](https://github.com/cardano-scaling/hydra/issues/2206) and [PR #2407](https://github.com/cardano-scaling/hydra/pull/2407)